Big Data/Kafka/Assignment1/requirements.txt
confluent-kafka=1.9.2=pypi_0
pyyaml=6.0=pypi_0
pandas=1.1.5=pypi_0
requests=2.28.1=pypi_0
jsonschema=4.16.0=pypi_0
